requirements.txt
attrs==24.2.0
cap==0.0.114
certifi==2024.8.30
charset-normalizer==3.4.0
clearml==1.16.5
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
deep-sort-realtime==1.3.2
enum34==1.1.10
filelock==3.13.1
filterpy==1.4.5
fonttools==4.54.1
fsspec==2024.2.0
furl==2.1.3
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
labelImg==1.8.6
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.2.1
nicer==0.0.36
numpy==1.26.3
opencv-python==4.10.0.84
orderedmultidict==1.0.1
packaging==24.1
pandas==2.2.3
pathlib2==2.3.7.post1
pillow==11.0.0
pkt==0.0.21
psutil==6.1.0
py-cpuinfo==9.0.0
PyJWT==2.8.0
pyparsing==3.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sympy==1.13.1
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tqdm==4.66.6
transforms==0.2.1
typing==3.7.4.3
typing_extensions==4.9.0
tzdata==2024.2
ultralytics==8.3.27
ultralytics-thop==2.0.10
urllib3==2.2.3
zipp==3.20.2
torch==2.3.1+cu121
torchaudio==2.3.1+cu121
torchvision==0.18.1+cu121

